+++ PracticoBiblioteca/Services/Implementaciones/UsuarioService.cs
@@ -17,31 +17,6 @@ public class UsuarioService : IUsuarioService
         _logger = logger;
     }
 
-    public async Task<SesionDTO?> AutenticacionAsync(LoginDTO login)
-    {
-        try
-        {
-            var response = await _httpClient.PostAsJsonAsync("api/usuarios/Authenticate", login);
-            response.EnsureSuccessStatusCode();
-
-            var result = await response.Content.ReadFromJsonAsync<SesionDTO>();
-            return result;
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError(ex, "Error en la autenticación");
-            throw;
-        }
-    }
-    public async Task<Usuario> RegistrarAsync(RegistroDTO dto)
-    {
-        var response = await _httpClient.PostAsJsonAsync("api/usuarios/register", dto);
-        response.EnsureSuccessStatusCode();
-
-        var usuario = await response.Content.ReadFromJsonAsync<Usuario>();
-        return usuario!;
-    }
-
     public async Task<List<Usuario>> ObtenerTodosAsync()
     {
         try
