+++ PracticoBiblioteca/Services/Interfaces/IUsuariosService.cs
@@ -6,8 +6,6 @@ namespace PracticoBiblioteca.Services.Interfaces;
 
 public interface IUsuarioService
 {
-    Task<SesionDTO?> AutenticacionAsync(LoginDTO login);
-    Task<Usuario> RegistrarAsync(RegistroDTO dto);
     Task<List<Usuario>> ObtenerTodosAsync();
     Task<Usuario?> ObtenerPorIdAsync(int id);
     Task AgregarAsync(Usuario usuario);
